Create LICENSE
From original README by minghai:

"All code is my original. Written in HTML + JavaScript. I declare the code in this web app as Public Domain. Only code, not images and sounds. Do what you want with my code. But I'm not responsible for anything, in any means."

Therefore the Unlicense is appropriate to this repository.